Refresh token

+++ src/utils/axios.js
@@ -29,10 +29,22 @@ axiosApiIntances.interceptors.response.use(
   function (error) {
     // Any status codes that falls outside the range of 2xx cause this function to trigger
     // Do something with response error
-    console.log(error.reponse);
     if (error.response.status === 403) {
-      localStorage.clear();
-      window.location.href = "login";
+      if (error.response.data.msg === "jwt expired") {
+        const refreshToken = localStorage.getItem("refreshToken");
+        console.log(refreshToken);
+        axiosApiIntances
+          .post("/refresh", { refreshToken })
+          .then((res) => {
+            console.log(res);
+          })
+          .catch((err) => {
+            console.log(err);
+          });
+      } else {
+        localStorage.clear();
+        window.location.href = "/login";
+      }
     }
     return Promise.reject(error);
   }
